requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
fastapi-cache2==0.2.2
redis==5.2.0
numpy==1.26.0
matplotlib==3.8.0
python-dotenv==1.0.0
pytest==7.4.2
pytest-asyncio==0.21.0
httpx==0.24.1
pydantic<2.0.0,>=1.6.2
requests==2.31.0
aiohttp==3.11.2
python-dotenv